feat(curried) Add curried functions (#19)
Add CurriedFunctions, which initializers as well as parameters, and are
formatted as `function(initializers)(parameters)`. These are accepted all the
same places a Function is allowed. A Function is now a CurriedFunction with no
initializers. This also makes the parameters optional, which is reflected by not
encoding the parameters if they are None.

There is a chunk of tech debt added here: the legacy API doesn't support curried
functions, they are sent as raw strings e.g. ["quantile(0.5)", "duration", None]
On top of this some users of the legacy API are sending functions in this form
as well e.g. ["apdex(300, duration)", None, None]. Snuba uses a regex grammar to
parse these strings, and it seemed overkill to port that to this SDK just for
this legacy use case. So instead the raw strings are passed through as is. Once
the legacy code is deprecated we can restrict this again.