feat(subquery) Add subquery support to SDK (#28)
Add the ability to pass another Query in the match clause of a query. Also add
validation for the inner query and to ensure the outer query is only using
fields returned by the inner query. Added a new visitor that can be used to find
all the expressions of a certain type in a query.