feat(join) Add Join/Relationship classes and query support (#35)
Add Join class that contains an array of Relationships, each of which describes
two entities and the join relationship between them. A query can now have a Join
as  the match value. This also adds a restriction that all columns in a Join
query must be qualified with an entity that also exists in the Join.