fix(groupby) Handle legacy queries with nested aliases of aggregates (#48)
This fix is to handle legacy queries where some of the expressions are functions
on aliases of other aggregate functions, e.g. `avg(x) AS avg_x, minus(avg_x, 2)`
When checking to see if the second function is an aggregate, also check all the
aggregate aliases found in the first pass of the function. I'm aware that this
will not work if aggregates are nested multiple levels down, but this will have
to do for now.  It also fixes a bug where if there is a non-aggregate function
on a column in the groupby, only the original column needs to be in the groupby,
 not the entire expression. E.g. `SELECT count(), toString(stuff) BY stuff`