fix(legacy) Handle weird edge case with escaping (#62)
There seems to be a bug in Sentry that sends search strings like this, but since
the legacy endpoint would allow this the SDK should make it work. Sometimes a
string is received with a single backslash on the end, which was incorrectly
escaping the single quote around the string and causing a parsing error. If
there is a trailing backslash, escape it.