fix(validate) Remove groupby checks (#70)
This code has been brittle and difficult to follow for a while, but what pushed
this change over the edge was queries being written for sessions. The `sessions`
column in Sentry is an alias for an aggregate function in Snuba, but the SDK
doesn't know that so it treats it as a non-aggregate and gets upset if it's not
included in the group by. So the SDK would need to know about all such aliases
in Snuba and it's just not worth it. Let Snuba handle this validation.