fix(legacy) Try to ensure tag conditions are on strings
Sometimes tag conditions are passed wrapped in functions. In particular with
datetimes, this behaviour in the legacy API was completely ambiguous, and Snuba
would leave the value as a string. In the SnQL translator, it would attempt to
convert the string to a datetime which is incorrect. In order to ensure that
SnQL remains backwards compatible, handle this case as well.