fix(legacy) Fix legacy nested functions
In very rare cases (only Sentry tests as far as I can tell) legacy functions are
nested ambiguously, without enclosing brackets to delineate the nested function.
The legacy parser had a special case to handle this. Import that functionality
to the legacy parser and use it when the original function fails.