ref(3.9) Upgrade snuba-sdk to use python 3.9
The sentry repository has been updated to 3.8, so we can now finally update and
drop support for 3.6. There is nothing preventing this library from using 3.9,
since it will be backwards compatible with 3.8 as well. This upgrade means that
__future__.annotations can be used.

Also update the required packages, which triggered a couple small formatting
changes:

- Correctly chain exceptions
- Follow PEP naming conventions for Exceptions