Add Entity Data Model to validate columns (#84)
Add a Data Model schema class that can be included in an Entity. A Data Model
specifies the columns that an Entity has and also any required columns and the
required time column. These values are used to validate queries to ensure that
all the required columns have conditions in the where clause.