feat: Add Lambdas and Identifiers expressions (#91)
Add Lambda and Identifier expressions to the syntax. Lambdas can be used to
build anonymous functions in the form (`x`) -> <transformation>. Identifiers
can be used inside the transformation to reference the identifier(s) in the
lambda. These can be used inside functions to support higher order functions
such as arrayMax.