feat: Add Request/Flags class instead of bare Query
This change will require Queries to be wrapped in a Request class before being
sent to Snuba. This class is used to maintain all the parts of the request
that are not in the SnQL language.

The Request has a dataset, AppID, Query and any Flags that are not part of the
SnQL specification. The dataset and AppID strings are required. The legacy
converter will automatically add an AppID if one isn't provided in order to
maintain backwards compatibility.

The Request looks like this:

[example]

Since the flags are now maintained by the Request class, they have been stripped
out of the query class and their logic simplified. This change also deprecates
the parent_api, team and feature flags. All of these have been replaced with the
AppID.

These changes should have no impact on what is actually sent to Snuba.