feat: Add parent_api flag to the Request (#106)
This flag has been useful in the past and we want to revisit it in the Snuba API
to help debug cases of abuse. Add it as an optional parameter so that this won't
break any existing uses of Request. This value should be updated automatically
in the snuba part of the sentry codebase.