feat(issue-platform): Add search_issues entity to `ENTITY_TIME_COLUMNS` (#108)
* feat(issue-platform): Add an entity time column for `search_issues`

Our search code uses the legacy method of querying snuba, so we'll need to add this here to make
date ranges in searches work.

* feat(issue-platform): Add `search_issues` entity to `ENTITY_TIME_COLUMNS`

Unfortunately issue search is still using the legacy query format for snuba. Adding this entity to
the legacy code so that it'll work in issue search.