feat(metrics): Add the base classes for MetricsQuery
This PR adds all the initial classes necessary to build a MetricsQuery. It also
changes Request to allow MetricsQuery objects.

There are some basic tests added as essentially placeholders. They are not
comprehensive. Also, this PR does not add any visitors necessary for
validation/serialization etc.

This change is not meant to be functional, it is meant to unblock other
development so that work can proceed on improving and using these classes.