feat(metrics) Add translators/validators for MetricsQuery
Add the basic SnQL translators for the MetricsQuery object and all the objects
it uses.

This adds a translator and validator visitor for the MetricsQuery class. It
also changes Timeseries, Metric, Rollup and MetricScope to all be in the same
file, and to not be Expressions. All of these classes return complex types when
being serialized, in that they affect multiple parts of the query separately.
For example, Rollup impacts the conditions as well as the ORDER BY clause. So
each of these classes also has a visitor in a similar way to the query.