ref(formula) Add SnQL visitor for Formula (#136)
First off, note that this is a temporary measure. Formulas don't encode nicely
into SnQL. The main issue with encoding MetricsQuery into SnQL is that the
visitor pattern assumes each component of a query can be encoded into a string,
and then combined together neatly into a query. That's not true of a
MetricsQuery, and especially a Formula, where the encoding of a single
component can impact multiple parts of a query. For example encoding a Metric
object influences the WHERE clause (metric_id filter) and the MATCH clause (for
the entity).

This assumption works much better when encoding MetricsQuery into MQL, since
the components can be encoded and combined together in a seamless way. Until
then, these visitors will have to hack around the restrictions.

This adds a simple SnQL visitor for the Formula class. This only works for the
most basic Formulas: no nested Formulas as allowed, and the groupby and entity
of any Timeseries parameters must match.