feat(mql) Support curried functions in MQL
One of the main uses for metrics is currently quantiles. Those are curried
functions, so MQL needs to support curried functions if it is going to be used
to encode queries being sent to Snuba.

Add curried functions to the MQL grammar that support arbitrary params. The
only use case at the moment is quantiles but it doesn't hurt to be a little
more open in the grammar.