ref(metrics): Support serializing a Request to MQL
Add a `serialize_to_mql` function to the Request class. This is a temporary
function that we can use as we cut over Sentry/Snuba to use the new MQL based
Snuba endpoint. Having two functions just makes it slightly easier to roll
back/forward, the calling code can be changed instead of having to release new
SDK versions or roll back versions.

Once Sentry/Snuba are using the new endpoint, this functionality can be folded
into the default `serialize` function in a future version.