ref(MQL): Simplify MQLContext validation/serialization (#149)
This arose from the realization that MQLContext is an internal class only. It
is only ever created when serializing a MetricsQuery, and that MetricsQuery
must be valid in order to be serialized in the first place.

Thus there's no need to do extra validation work of the MQLContext. Once that
was removed, the serializer also didn't seem necessary. The objects themselves
could be serialized and simply joined together in the MQLContext class.

The class itself is still useful, to provide a human readable understanding of
what fields are allowed in a serialized MQL query.

This also fixes a bug with the indexer_mappings field. That field will need to
be passed from Sentry to Snuba, in order for Snuba to do resolving. That was
added to the MetricsQuery class. In the future the resolving itself will happen
in Snuba, but for now this is required.