feat(mql): Support encoding Formulas to MQL (#154)
* feat(mql): Support encoding Formulas to MQL

This PR adds a visitor that will encode a Formula to MQL. It supports nested
Formulas as well.

Also, even though the parser supported parsing Formulas from MQL, the tests
were all disabled. The tests were enabled, and used as a basis for all the
other test for serializing Formulas.

For testing, when a Formula is serialized to MQL, the test also checks that the
output MQL can be deserialized back to an SDK object.

An explicit decision was made to not do reverse testing when deserializing MQL.
That's because the range of possible inputs that can be correctly deserialized
is much broader than the set of possible outputs from serializing. For example,
`((((((sum(foo)))))))` is valid input, but when that is serialized it won't
match.

* add CHANGES

* remove comments

* add test for OR conditions