feat: Support MQL strings in MetricsQuery
Allow providing a MQL string as the `query` parameter to a MetricsQuery. The
string will get parsed and validated when being serialized to ensure that it's
a valid MQL string.

Since an MQL string can't specify an entity, this requires that Snuba can
automatically infer the entity based on the passed in MRI.